+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- Child svg is now removed from the page flow with 'position: absolute;'.
+- Removed unnecessary svg properties xmlns and version.
 
 ## [1.4.0] - 2024-11-19
 ### Added